[Bronze V] Title: 두 수 비교하기, Time: 124 ms, Memory: 14224 KB -BaekjoonHub

+++ 백준/Bronze/1330. 두 수 비교하기/두 수 비교하기.java
@@ -8,8 +8,8 @@ public class Main {
 
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        long A = Long.parseLong(st.nextToken());
-        long B = Long.parseLong(st.nextToken());
+        int A = Integer.parseInt(st.nextToken());
+        int B = Integer.parseInt(st.nextToken());
 
         if (A > B) {
             bw.write(">");
